Implemented full async mode based on "aiohttp" library
