+++ custom_components/smartthinq_sensors/const.py
@@ -2,7 +2,7 @@
 Support to interface with LGE ThinQ Devices.
 """
 
-__version__ = "0.13.3"
+__version__ = "0.20.0"
 PROJECT_URL = "https://github.com/ollo69/ha-smartthinq-sensors/"
 ISSUE_URL = "{}issues".format(PROJECT_URL)
 
@@ -15,8 +15,6 @@ __min_ha_version__ = f"{MIN_HA_MAJ_VER}.{MIN_HA_MIN_VER}.0"
 CONF_LANGUAGE = "language"
 CONF_OAUTH_URL = "outh_url"
 CONF_USE_API_V2 = "use_api_v2"
-CONF_USE_TLS_V1 = "use_tls_v1"
-CONF_EXCLUDE_DH = "exclude_dh"
 
 CLIENT = "client"
 LGE_DEVICES = "lge_devices"
