+++ custom_components/smartthinq_sensors/manifest.json
@@ -8,5 +8,5 @@
   "requirements": ["pycountry>=20.7.3", "xmltodict>=0.12.0"],
   "config_flow": true,
   "iot_class": "cloud_polling",
-  "version": "0.13.3"
+  "version": "0.20.0"
 }
