+++ custom_components/smartthinq_sensors/translations/en.json
@@ -3,7 +3,8 @@
         "abort": {
             "single_instance_allowed": "Only a single configuration of SmartThinQ LGE Sensors is allowed.",
             "no_smartthinq_devices": "No SmartThinQ devices found. Component setup aborted.",
-            "unsupported_version": "This integration require at least HomeAssistant version {req_ver}, you are running version {run_ver}."
+            "unsupported_version": "This integration require at least HomeAssistant version {req_ver}, you are running version {run_ver}.",
+            "reconfigured": "Configuration successfully completed."
         },
         "error": {
             "error_url": "Error retrieving login URL from ThinQ.",
@@ -11,6 +12,7 @@
             "invalid_language": "Invalid language format.",
             "invalid_url": "Invalid redirection URL. Make sure you can access the ThinQ app on your phone. You can get support [over here](https://git.io/JTjid).",
             "invalid_credentials": "Invalid SmartThinQ credentials. Try again, and make sure you can access the ThinQ app on your phone.",
+            "invalid_config": "Found invalid configuration, please reconfigure.",
             "no_user_info": "User Name and Password are required.",
             "unknown": "Unknown error."
         },
@@ -25,7 +27,7 @@
                     "use_tls_v1": "Force use of protocol TLSv1 (sometimes required)",
                     "exclude_dh": "Do not use DH cypher (in case 'dh key too small' error)"
                 },
-                "description": "Select your SmartThinQ account localization information.",
+                "description": "Insert your SmartThinQ account access information.",
                 "title": "SmartThinQ LGE Sensors"
             },
             "url": {
