+++ custom_components/smartthinq_sensors/translations/it.json
@@ -3,7 +3,8 @@
         "abort": {
             "single_instance_allowed": "E' consentita una sola configurazione di SmartThinQ LGE Sensors.",
             "no_smartthinq_devices": "Nessun dispositivo SmartThinQ trovato. Component setup interrotto.",
-            "unsupported_version": "Questa integrazione richiede almeno la versione {req_ver} di HomeAssistant, tu stai usando la versione {run_ver}."
+            "unsupported_version": "Questa integrazione richiede almeno la versione {req_ver} di HomeAssistant, tu stai usando la versione {run_ver}.",
+            "reconfigured": "Configurazione completata con successo."
         },
         "error": {
             "error_url": "Errore nel recuperare l'URL di login da ThinQ.",
@@ -11,6 +12,7 @@
             "invalid_language": "Formato lingua non valido.",
             "invalid_url": "URL di reindizzamento non valido. Accertati di poter accedere alla app ThinQ sul tuo smartphone. Puoi richiedere supporto [qui](https://git.io/JTjid).",
             "invalid_credentials": "Credenziali SmartThinQ non valide. Riprova, assicurandoti prima di poter accedere dall'applicazione ThinQ sul telefono",
+            "invalid_config": "Trovata configurazione non valida, riconfigurare.",
             "no_user_info": "Nome utente e Password sono campi obbligatori.",
             "unknown": "Errore sconosciuto."
         },
@@ -25,7 +27,7 @@
                     "use_tls_v1": "Forza l'utilizzo del protocollo TLSv1 (richiesto in alcuni casi)",
                     "exclude_dh": "Non utilizzare il cypher DH (in caso di errore 'dh key too small')"
                 },
-                "description": "Seleziona le informazioni di localizzazione del tuo account SmartThinQ.",
+                "description": "Imposta le informazioni di accesso del tuo account SmartThinQ.",
                 "title": "SmartThinQ LGE Sensors"
             },
             "url": {
